CalloutContent: Convert CalloutContent to JS Styling (#3745)
* Add base callout file

* Rename file to CalloutContent.base

* CalloutContentbase

* Callout content styling

* Callout content styling

* Add hostid

* Revert "Add hostid"

This reverts commit d5a842d8ebc099fa4652d7abd7205bdfcf033b0a.

* Animation styles

* Refactor base callout content to use js styling

* Create calloutContent const

* Fix import

* Styling changes

* FIx beak background styling and boxShadow

* Fix callout width

* Add inline styling from popup

* Remove callout.scss file

* Add change file

* Update snapshot and maxheight and masfontsize to CSSPixelUnitRule

* Add focus clear mixin to styling package an use it

* Add remaining change files

* Add base callout file

* Rename file to CalloutContent.base

* CalloutContentbase

* Callout content styling

* Callout content styling

* Animation styles

* Refactor base callout content to use js styling

* Create calloutContent const

* Fix import

* Styling changes

* FIx beak background styling and boxShadow

* Fix callout width

* Add inline styling from popup

* Remove callout.scss file

* Add change file

* Update snapshot and maxheight and masfontsize to CSSPixelUnitRule

* Add focus clear mixin to styling package an use it

* Add remaining change files

* Add comments

* Add comment for beak style

* Test using CalloutContent

* Use spread operator
