+++ common/changes/@uifabric/merge-styles/malozano-CalloutStyling_2018-01-19-02-15.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@uifabric/merge-styles",
+      "comment": "Change maxHeight/maxFontSize to CSSPixelUnitRule (number and string)",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@uifabric/merge-styles",
+  "email": "malozano@microsoft.com"
+}
