+++ common/changes/@uifabric/styling/malozano-CalloutStyling_2018-01-19-02-15.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "packageName": "@uifabric/styling",
+      "comment": "Add focus clear func to styling and change maxheight/maxfontsize to CSSPixelUnitRule ",
+      "type": "minor"
+    }
+  ],
+  "packageName": "@uifabric/styling",
+  "email": "malozano@microsoft.com"
+}
