--- packages/office-ui-fabric-react/src/components/Callout/Callout.scss
@@ -1,63 +1,0 @@
-@import '../../common/common';
-
-// Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE in the project root for license information.
-
-//
-// Office UI Fabric
-// --------------------------------------------------
-// Callout styles
-
-$Callout-beak-slant-width: 28px;
-$Callout-smallbeak-slant-width: 16px;
-
-@mixin calloutBeak {
-  position: absolute;
-  background-color: $ms-color-white;
-  box-shadow:  inherit;
-  border: inherit;
-  box-sizing: border-box;
-  -webkit-transform: rotate(45deg);
-  -ms-transform: rotate(45deg);
-  transform: rotate(45deg);
-}
-
-.root {
-  @include ms-drop-shadow(0,0, 5px, 0px);
-  position: absolute;
-  box-sizing: border-box;
-  border: 1px solid $ms-color-neutralLight;
-
-  @include focus-clear();
-
-  @include high-contrast {
-    border: 1px solid WindowText;
-  }
-}
-
-.container {
-  position: relative;
-}
-
-// The actual callout element
-.main {
-  background-color: $ms-color-white;
-  overflow-x: hidden;
-  position: relative;
-}
-
-.overFlowYHidden {
-  overflow-y: hidden;
-}
-
-.beak {
-  @include calloutBeak;
-}
-
-.beakCurtain {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  background-color: $ms-color-white;
-}
