+++ packages/office-ui-fabric-react/src/components/Callout/Callout.tsx
@@ -4,7 +4,8 @@ import * as React from 'react';
 
 import { BaseComponent } from '../../Utilities';
 import { ICalloutProps } from './Callout.types';
-import { ICalloutState, CalloutContent } from './CalloutContent';
+import { ICalloutState } from './CalloutContent.base';
+import { CalloutContent } from './CalloutContent';
 import { Layer } from '../../Layer';
 
 export class Callout extends BaseComponent<ICalloutProps, ICalloutState> {
