+++ packages/office-ui-fabric-react/src/components/Callout/Callout.types.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { Callout } from './Callout';
+import { IStyle, ITheme } from '../../Styling';
 import { CalloutContent } from './CalloutContent';
 import { DirectionalHint } from '../../common/DirectionalHint';
 import {
@@ -7,12 +8,13 @@ import {
   IRectangle
 } from '../../Utilities';
 import { ICalloutPositionedInfo } from '../../utilities/positioning';
+import { IStyleFunction } from '../../Utilities';
 
 export interface ICallout {
 
 }
 
-export interface ICalloutProps extends React.Props<Callout | CalloutContent> {
+export interface ICalloutProps {
   /**
    * Optional callback to access the ICallout interface. Use this instead of ref for accessing
    * the public methods and properties of the component.
@@ -192,4 +194,90 @@ export interface ICalloutProps extends React.Props<Callout | CalloutContent> {
    * Callback when the Callout body is scrolled.
    */
   onScroll?: () => void;
+
+  /**
+   * Optional theme for component
+   */
+  theme?: ITheme;
+
+  /**
+  * Optional styles for the component.
+  */
+  getStyles?: IStyleFunction<ICalloutContentStyleProps, ICalloutContentStyles>;
+}
+
+export interface ICalloutContentStyleProps {
+  /**
+   * Theme to apply to the calloutContent.
+   */
+  theme: ITheme;
+
+  /**
+   * Width for callout including borders.
+   */
+  calloutWidth?: number;
+
+  /**
+ * CSS class to apply to the callout.
+ */
+  className?: string;
+
+  /**
+   * Callout positioning data
+   */
+  positions?: ICalloutPositionedInfo;
+
+  /**
+   * Whether or not to clip content of the callout,
+   * if it overflows vertically.
+   */
+  overflowYHidden?: boolean;
+
+  /**
+   * Max height applied to the content of a callout.
+   */
+  contentMaxHeight?: number;
+
+  /**
+   * Background color for the beak and callout.
+   */
+  backgroundColor?: string;
+
+  /**
+   * Width of Callout beak
+   */
+  beakWidth?: number;
+
+  /**
+    * Deprecated at v0.59.1, to be removed at >= v1.0.0. Pass in a beakWidth to dictate size.
+    * @deprecated
+    */
+  beakStyle?: string;
+}
+
+export interface ICalloutContentStyles {
+  /**
+   * Style for wrapper of Callout component.
+   */
+  container: IStyle;
+
+  /**
+  * Style for callout container root element.
+  */
+  root: IStyle;
+
+  /**
+  * Style for callout beak.
+  */
+  beak: IStyle;
+
+  /**
+  * Style for callout beak curtain.
+  */
+  beakCurtain: IStyle;
+
+  /**
+  * Style for content component of the callout.
+  */
+  calloutMain: IStyle;
 }
