+++ packages/styling/src/styles/index.ts
@@ -2,7 +2,7 @@ export { AnimationStyles, AnimationVariables } from './AnimationStyles';
 export { DefaultPalette } from './DefaultPalette';
 export { DefaultFontStyles, registerDefaultFontFaces } from './DefaultFontStyles';
 export { FontSizes, FontWeights, IconFontSizes, createFontStyles } from './fonts';
-export { getFocusStyle } from './getFocusStyle';
+export { getFocusStyle, focusClear } from './getFocusStyle';
 export { hiddenContentStyle } from './hiddenContentStyle';
 export {
   ThemeSettingName,
